initialise Prisma Client
