+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@prisma/client": "^4.8.1",
     "bcrypt": "^5.0.1",
     "chalk": "^5.0.0",
     "cheerio": "^1.0.0-rc.10",
